# tests/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "admin-on-rest entity APIs"
  },
  "host": "localhost:7000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
...
  "securityDefinitions": {
    "OAuth2": {
      "type": "oauth2",
      "flow": "password",
      "tokenUrl": "https://info.services.auth.localhost/oauth2/tokeninfo",
      "scopes": {
        "uid": "Unique identifier of the user accessing the service."
      }
    },
    "BearerAuth": {
      "description": "Bearer: xxxxxx.yyyyyyy.zzzzzz\n",
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    },
    "jwt": {
      "type": "apiKey",
      "name": "Authorization",
      "description": "JWT bearer token",
      "in": "header"
    }
  },
  "security": [
    {
      "key": [

      ]
    },
    {
      "OAuth2": [
        "uid"
      ]
    },
    {
      "BearerAuth": [

      ]
    }
  ],
  "paths": {
...
    "/role/_roles": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "summary": "list",
        "operationId": "getRoles",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Role"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "options": {
        "tags": [
          "role-controller"
        ],
        "summary": "addRole",
        "operationId": "addRoleUsingOPTIONS",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "role",
            "description": "role",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "role-controller"
        ],
        "summary": "addRole",
        "operationId": "addRoleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "role",
            "description": "role",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/schemas/_entitys/{eid}": {
      "get": {
        "tags": [
          "schema-controller"
        ],
        "summary": "findOne",
        "operationId": "findSchemaEntityGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "eid",
            "in": "path",
            "description": "eid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Entity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/user/_users": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "list",
        "operationId": "listUsersGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "options": {
        "tags": [
          "user-controller"
        ],
        "summary": "addUser",
        "operationId": "addUserUsingOPTIONS",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "addUser",
        "operationId": "addUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Entity": {
      "type": "object",
      "properties": {
        "crud": {
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "c",
              "r",
              "u",
              "d"
            ]
          }
        },
        "fields": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Field"
          }
        },
        "id": {
          "type": "string"
        },
        "label": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "redirect": {
          "type": "string",
          "enum": [
            "edit",
            "show",
            "list"
          ]
        },
        "showInMenu": {
          "type": "boolean"
        }
      }
    },
    "Field": {
      "type": "object",
      "properties": {
        "choices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/IChoiceItem"
          }
        },
        "component": {
          "type": "string",
          "enum": [
            "Autocomplete",
            "Boolean",
            "NullableBoolean",
            "CheckboxGroup",
            "Date",
            "Disabled",
            "File",
            "Image",
            "LongText",
            "Number",
            "RadioButtonGroup",
            "ReferenceArray",
            "Reference",
            "RichText",
            "SelectArray",
            "Select",
            "Text"
          ]
        },
        "dataSourceId": {
          "type": "string"
        },
        "defaultValue": {
          "type": "string"
        },
        "eid": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "inputType": {
          "type": "string",
          "enum": [
            "text",
            "email",
            "password",
            "url"
          ]
        },
        "isAutoIncremented": {
          "type": "boolean"
        },
        "isPartOfPrimaryKey": {
          "type": "boolean"
        },
        "label": {
          "type": "string"
        },
        "maxLength": {
          "type": "integer",
          "format": "int32"
        },
        "maxValue": {
          "type": "string"
        },
        "minValue": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "reference": {
          "type": "string"
        },
        "referenceOptionText": {
          "type": "string"
        },
        "required": {
          "type": "boolean"
        },
        "showInCreate": {
          "type": "boolean"
        },
        "showInEdit": {
          "type": "boolean"
        },
        "showInFilter": {
          "type": "boolean"
        },
        "showInList": {
          "type": "boolean"
        },
        "showInShow": {
          "type": "boolean"
        }
      }
    },
...
    "IChoiceItem": {
      "type": "object"
    },
...
    "Role": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "id": {
          "type": "string"
        },
        "lastPasswordResetDate": {
          "type": "string",
          "format": "date-time"
        },
        "password": {
          "type": "string"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          }
        },
        "username": {
          "type": "string"
        }
      }
    }
  }
}